fix: Rename socket-listeners.ts to socketListeners.ts - ADR-011 compliance
ADR-011 enforces camelCase for file names. Renamed:
  packages/frontend/src/services/socket-listeners.ts → socketListeners.ts

Updated imports:
  packages/frontend/src/hooks/useWebSocketCache.ts

Ensures naming consistency with project conventions.
